+++ CFunctionDefinitionNode.cpp
@@ -43,13 +43,13 @@ void	CFunctionDefinitionNode::AddLocalVar( const std::string& inName,
 }
 
 
-long	CFunctionDefinitionNode::GetBPRelativeOffsetForLocalVar( const std::string& inName )
+int16_t	CFunctionDefinitionNode::GetBPRelativeOffsetForLocalVar( const std::string& inName )
 {
 	std::map<std::string,CVariableEntry>::iterator	foundVariable = mLocals.find( inName );
 	if( foundVariable != mLocals.end() )
 	{
-		long	bpRelOffs = foundVariable->second.mBPRelativeOffset;
-		if( bpRelOffs == LONG_MAX )
+		int16_t	bpRelOffs = foundVariable->second.mBPRelativeOffset;
+		if( bpRelOffs == INT16_MAX )
 		{
 			bpRelOffs = mLocalVariableCount++;
 			foundVariable->second.mBPRelativeOffset = bpRelOffs;
@@ -57,7 +57,7 @@ long	CFunctionDefinitionNode::GetBPRelativeOffsetForLocalVar( const std::string&
 		return bpRelOffs;
 	}
 	else
-		return LONG_MAX;
+		return INT16_MAX;
 }
